+++ romancal/pipeline/__init__.py
@@ -1,0 +1,7 @@
+"""
+This module collects all of the stpipe.Pipeline subclasses
+made available by this package.
+"""
+
+
+__all__ = []