+++ CHANGES.rst
@@ -1,6 +1,12 @@
 0.2.0 (Unreleased)
 ==================
 
+stpipe
+------
+
+- Create stpipe module which provides Roman-specific Step and Pipeline
+  subclasses. [#103].
+
 0.1.0 (2020-12-11)
 ==================
 
@@ -8,8 +14,7 @@ datamodels
 ----------
 
 - First release of romancal. Includes the core metadata and a ``FlatModel``.
-  
+
 - Update date strings in schemas and tests from strings to astropy objects [#32]
-  
+
 -  Update Flat Schema for DQ Array DType [#55]
-